gradient: smooth + smooth, test on trajectories with one collision on 4 edges success

+++ program/test_params.py
@@ -34,7 +34,7 @@ def plot_trajectory(params, trajectories, epoch=0, writer=None):
     for trajectory_index, data_set in enumerate(trajectories):
         # trajectory_index = index  # choose which trajectory to test, current total 150 trajectories 2022.06.21
         init_state = calculate_init_state(data_set).cpu().numpy()
-        state_num = (data_set[-1, -1] - data_set[0, -1]) * 120
+        state_num = int((data_set[-1, -1] - data_set[0, -1]) * 120)
         table = air_hockey_baseline.AirHockeyTable(length=1.948, width=1.038, goalWidth=0.25, puckRadius=0.03165,
                                                    restitution=params[2], rimFriction=params[3], dt=1 / 120)
         system = air_hockey_baseline.SystemModel(tableDamping=params[1], tableFriction=params[0], tableLength=1.948,
